+++ R/add_hier_approach.R
@@ -4,7 +4,8 @@ NULL
 #' Add a hierarchical approach
 #'
 #' Add a hierarchical (lexicographic) approach for multi-objective optimization
-#' to a multi-objective conservation planning problem (Jaimes *et al.* 2009).
+#' to a multi-objective conservation planning problem
+#' (López Jaimes *et al.* 2009).
 #' Broadly speaking, this approach involves using multiple optimization
 #' procedures to solve each [problem()] in a
 #' [multi_problem()] object following a hierarchical (lexicographic) ordering,
@@ -231,7 +232,8 @@ NULL
 #' sums of objectives for Pareto set generation in multicriteria optimization
 #' problems. _Structural Optimization_, **14**: 63--69.
 #'
-#' Jaimes AL, Saúl ZM, and Coello Coello CA (2009) *An introduction to
+#' López Jaimes A, Zapotecas Martínez S, and Coello Coello CA (2009)
+#' *An introduction to
 #' multiobjective optimization techniques* in Optimization in Polymer
 #' Processing. Eds Gaspar-Cunha A and Covas JA. Nova Science Publishers Inc,
 #' New York, United States.