+++ R/multi_problem.R
@@ -73,6 +73,12 @@ NULL
 #' planning to meet global biodiversity goals.
 #' _Trends in Ecology and Evolution_, **40**: 395--410.
 #'
+#' Neubert S, McGowan J, Metcalfe K, Hanson JO, Buenafe KCV, Dabalà A, Dunn DC,
+#' Everett JD, Possingham HP, Stelzenmüller V, Estep A, Ervin J, and
+#' Richardson AJ (2025) Multiple-use spatial planning for sustainable
+#' development and conservation. _Trends in Ecology and Evolution_, **40**:
+#' 1126--1142.
+#'
 #' Williams PJ and Kendall WL (2017) A guide to multi-objective optimization
 #' for ecological problems with an application to cackling goose management.
 #' _Ecological Modelling_, **343**: 54-67.